Test 3 branch (#70)
* Refactored: Moved tests from Rmd files to 'tests' directory

* Refactored code to generate files in 'exercises' directory instead of 'tests'

* Refactored code to extract common functionality into a new file and implemented cleanup for test-created files.

* made a new branch

* fixed the typo in the file name

* fixed the typo in the file name

* Corrected typos and restructured test file organization.

* Fix spacing issue, correct grammar mistake, and include sampe_data.csv in repository.

* Add test files for exercises in chapters 10, 11, 12, 14, and 15

* Reverted file to original state without solutions

* removed data from git repository

* Fixed issue with extra curly braces in chapter 7 exercises

* fixed the file address

* Added purl=False to chapters 11, 12, 14, and 15

* Fixed typo in purl=False

* removed printing test_directories in run_all_tests.R

* Fixed incorrect address

* Corrected typos across chapters 9 to 15

* Fix file existence check approach

* Refactored cleanup process for improved clarity

* Fixed typos, cross-references duplication, and added library imports at the beginning of the exercises

* Removed unused CSV file

* Refactor Chapter 13 exercises to utilize downloaded file from Shiny app

* Fix inconsistent spacing

* Fix further inconsistent spacing

* Fixed grammar and spelling errors in chapters 16-21; added test files for chapters 19 and 20.

* Fixed spelling/grammar errors; updated Chapter 19 Test 4; moved Going Deeper to main text

* Resolved further merge conflicts

* Ensure consistent titles for subsections in chapter 16  and correct typo

* Improve error messages in exercise test files (excluding Rmd files)

* Fix incorrect file path in Chapter 19 tests

* Fix backslash issues inside chapter7.Rmd

* unused helper files removed

+++ exercises/chapter 7/chapter7_tests.R
@@ -1,39 +1,80 @@
 #Unit tests for chapter 7 exercises
 
-# Load  packages
+# Load packages
 library(testthat)
 library(here)
 library(fs)
+library(withr)
+library(knitr)
+library(tidyverse)
 
-
-if (file.exists("exercises/helper_functions.R")) {
-  source("exercises/helper_functions.R")
+# Define the function to run all chunks from Rmd file
+run_all_chunks <- function(rmd, envir = globalenv()) {
+  tempR <- file.path("temp.R")  # Create temporary file in the same directory
+  on.exit(unlink(tempR))
+  
+  purl(input = rmd, output = tempR)
+  
+  # Source the temporary R script with chdir = TRUE so that the working directory is set to the Rmd file's directory
+  source(tempR, chdir = TRUE, local = envir)
+  
 }
 
-# Specify the path to your Rmd file using here::here()
-rmd_file_path <- here("exercises", "chapter 7", "chapter7.Rmd")
+# Function to determine the correct path for the Rmd file
+get_rmd_path <- function(folder_name, filename) {
+  paths <- list(here(filename), here("exercises", folder_name, filename))
+  for (path in paths) {
+    if (file.exists(path))
+      return(path)
+  }
+  stop("Rmd file not found!")
+}
 
-# Exercise 1: Installing Packages
-test_that("Installing Packages", {
-  tmp_dir <- tempdir()
-  # copy contents of current directory to temp directory
-  dir_copy(dirname(rmd_file_path), tmp_dir, overwrite = TRUE)
-  
-  withr::with_dir(tempdir(), {
-    # Run all chunks in the Rmd file to execute the exercises
-    suppressWarnings(run_all_chunks(rmd_file_path))
-    # Task 1: Check if xgboost is installed
-    expect_true("xgboost" %in% installed.packages()[, "Package"])
+check_type <- function(obj) {
+  if (is.list(obj) && !is.data.frame(obj)) {
+    return("list")
+  } else if (is.data.frame(obj)) {
+    return("data.frame")
+  } else if (is.matrix(obj)) {
+    return("matrix")
+  } else if (is.vector(obj)) {
+    return(paste0("vector (", typeof(obj), ")"))
+  } else if (is.factor(obj)) {
+    return("factor")
+  } else if (is.array(obj)) {
+    return("array")
+  } else {
+    return(class(obj)[1])  # Returns the first class if multiple
+  }
+}
+
+# Specify the path to your Rmd file
+rmd_file_path <- get_rmd_path("chapter 7", "chapter7.Rmd")
+
+#This test can be used only if it is run locally
+if (rmd_file_path == here("exercises", "chapter 7", "chapter7.Rmd")) {
+  # Exercise 1: Installing Packages
+  test_that("Installing Packages", {
+    tmp_dir <- tempdir()
+    # copy contents of current directory to temp directory
+    dir_copy(dirname(rmd_file_path), tmp_dir, overwrite = TRUE)
     
-    # Task 2: Check if XML and mlr3 are installed
-    packages_installed <- c("XML", "mlr3")
-    expect_true(all(packages_installed %in% installed.packages()[, "Package"]))
+    with_dir(tempdir(), {
+      # Run all chunks in the Rmd file to execute the exercises
+      suppressWarnings(run_all_chunks(rmd_file_path))
+      # Task 1: Check if xgboost is installed
+      expect_true("xgboost" %in% installed.packages()[, "Package"])
+      
+      # Task 2: Check if XML and mlr3 are installed
+      packages_installed <- c("XML", "mlr3")
+      expect_true(all(packages_installed %in% installed.packages()[, "Package"]))
+    })
+    
+    # remove the temp directory
+    unlink(tmp_dir, recursive = TRUE)
+    
   })
-  
-  # remove the temp directory
-  unlink(tmp_dir, recursive = TRUE)
-  
-})
+}
 
 # Exercise 2: Reading a CSV file -- tidyverse
 test_that("Reading a CSV File with readr", {
@@ -42,10 +83,27 @@ test_that("Reading a CSV File with readr", {
   # copy contents of current directory to temp directory
   dir_copy(dirname(rmd_file_path), tmp_dir, overwrite = TRUE)
   
-  withr::with_dir(tempdir(), {
+  with_dir(tempdir(), {
     # Run all chunks in the Rmd file to execute the exercises
     suppressWarnings(run_all_chunks(rmd_file_path))
-    expect_true(is.data.frame(csv_data) && nrow(csv_data) > 0)
+    # Check if csv_data is a data frame
+    expect_equal(
+      is.data.frame(csv_data),
+      TRUE,
+      info = sprintf(
+        "Expected csv_data to be a data frame. It was: %s",
+        check_type(csv_data)
+      )
+    )
+    # Check if csv_data has more than 0 rows
+    expect_equal(
+      nrow(csv_data) > 0,
+      TRUE,
+      info = sprintf(
+        "Expected csv_data to have more than 0 rows. It has: %d rows",
+        nrow(csv_data)
+      )
+    )
   })
   
   # remove the temp directory
@@ -59,11 +117,30 @@ test_that("Scatterplot with ggplot2", {
   # copy contents of current directory to temp directory
   dir_copy(dirname(rmd_file_path), tmp_dir, overwrite = TRUE)
   
-  withr::with_dir(tempdir(), {
+  with_dir(tempdir(), {
     # Run all chunks in the Rmd file to execute the exercises
     suppressWarnings(run_all_chunks(rmd_file_path))
-    expect_true(inherits(scatter_plot, "ggplot") &&
-                  length(scatter_plot$layers) > 0)
+    
+    # Check if scatter_plot is a ggplot object
+    expect_equal(
+      inherits(scatter_plot, "ggplot"),
+      TRUE,
+      info = sprintf(
+        "Expected scatter_plot to be a ggplot object. Actual class: %s",
+        class(scatter_plot)
+      )
+    )
+    
+    # Check if scatter_plot has at least one layer
+    expect_equal(
+      length(scatter_plot$layers) > 0,
+      TRUE,
+      info = sprintf(
+        "Expected scatter_plot to have at least one layer. The number of layers is: %d",
+        length(scatter_plot$layers)
+      )
+    )
+    
   })
   
   # remove the temp directory
@@ -77,18 +154,70 @@ test_that("Advanced filtering and summarization", {
   # copy contents of current directory to temp directory
   dir_copy(dirname(rmd_file_path), tmp_dir, overwrite = TRUE)
   
-  withr::with_dir(tempdir(), {
+  with_dir(tempdir(), {
     # Run all chunks in the Rmd file to execute the exercises
     suppressWarnings(run_all_chunks(rmd_file_path))
-    expect_true(all(c("mean_score", "sd_score") %in% names(stats_result)))
-    expect_equal(nrow(stats_result), 1)
-    expect_true(
-      is.numeric(stats_result$mean_score) &&
-        stats_result$mean_score >= 0 &&
-        stats_result$mean_score <= 100
+    
+    # Check if the required columns are present in stats_result
+    expected_columns <- c("mean_score", "sd_score")
+    missing_columns <- setdiff(expected_columns, names(stats_result))
+    expect_equal(
+      length(missing_columns) == 0,
+      TRUE,
+      info = sprintf(
+        "Missing columns in stats_result: %s",
+        paste(missing_columns, collapse = ", ")
+      )
     )
-    expect_true(is.numeric(stats_result$sd_score) &&
-                  stats_result$sd_score >= 0)
+    
+    # Check if stats_result has exactly 1 row
+    expect_equal(
+      nrow(stats_result),
+      1,
+      info = sprintf(
+        "Expected stats_result to have 1 row. Actual number of rows: %d",
+        nrow(stats_result)
+      )
+    )
+    
+    # Check if mean_score is numeric
+    expect_equal(
+      is.numeric(stats_result$mean_score),
+      TRUE,
+      info = sprintf(
+        "Expected mean_score to be numeric. Actual class: %s",
+        class(stats_result$mean_score)
+      )
+    )
+    
+    # Check if mean_score is in range [0, 100]
+    expect_equal(
+      stats_result$mean_score >= 0 &&
+        stats_result$mean_score <= 100,
+      TRUE,
+      info = sprintf(
+        "Expected mean_score to be between 0 and 100. Actual value: %s",
+        toString(stats_result$mean_score)
+      )
+    )
+    # Check if sd_score is numeric
+    expect_equal(
+      is.numeric(stats_result$sd_score),
+      TRUE,
+      info = sprintf(
+        "Expected sd_score to be numeric. Actual class: %s",
+        class(stats_result$sd_score)
+      )
+    )
+    # check if sd_score is non-negative
+    expect_equal(
+      stats_result$sd_score >= 0,
+      TRUE,
+      info = sprintf(
+        "Expected sd_score to be non-negative. Actual value: %s",
+        toString(stats_result$sd_score)
+      )
+    )
   })
   
   # remove the temp directory
